lambda-validation/hello_world/requirements.txt
requests
google-api-python-client
pandas
snowflake-connector-python
s3fs